Integrated HTTP requests with a repository via event mechanism (#10)

+++ src/main/java/uk/co/mulecode/fileservice/repository/dto/HttpRequestEntity.java
@@ -43,5 +43,5 @@ public class HttpRequestEntity implements Serializable {
     private String requestIpProvider;
     @NotNull
     @Column(name = "timeLapsed")
-    private int timeLapsed;
+    private long timeLapsed;
 }
